+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project (loosely) adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+* Metadata asset field
 
 ## [1.2.0](https://github.com/animavirtuality/onlybots-core/compare/1.1.2...1.2.0) - October 7, 2022
 ### Changed