+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project (loosely) adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [2.0.1](https://github.com/animavirtuality/onlybots-core/compare/2.0.0...2.0.1) - October 20, 2022
 ### Fixed
 * Fix bug when all layers are compressed as fields
 * Fix bug with list compression